修订com.dlink.ud更换为com.dlink.function
