--- dlink-function/src/main/java/com/dlink/ud/udtaf/Top2.java
@@ -1,74 +1,0 @@
-/*
- *
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.dlink.ud.udtaf;
-
-import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.table.functions.TableAggregateFunction;
-import org.apache.flink.util.Collector;
-
-/**
- * 官网Demo Top2
- *
- * @author wenmo
- * @since 2021/6/14 20:44
- */
-
-public class Top2 extends TableAggregateFunction<Tuple2<Integer, Integer>, Top2.Top2Accumulator> {
-
-    public static class Top2Accumulator {
-        public Integer first;
-        public Integer second;
-    }
-
-    @Override
-    public Top2Accumulator createAccumulator() {
-        Top2Accumulator acc = new Top2Accumulator();
-        acc.first = Integer.MIN_VALUE;
-        acc.second = Integer.MIN_VALUE;
-        return acc;
-    }
-
-    public void accumulate(Top2Accumulator acc, Integer value) {
-        if (value > acc.first) {
-            acc.second = acc.first;
-            acc.first = value;
-        } else if (value > acc.second) {
-            acc.second = value;
-        }
-    }
-
-    public void merge(Top2Accumulator acc, Iterable<Top2Accumulator> it) {
-        for (Top2Accumulator otherAcc : it) {
-            accumulate(acc, otherAcc.first);
-            accumulate(acc, otherAcc.second);
-        }
-    }
-
-    public void emitValue(Top2Accumulator acc, Collector<Tuple2<Integer, Integer>> out) {
-        // emit the value and rank
-        if (acc.first != Integer.MIN_VALUE) {
-            out.collect(Tuple2.of(acc.first, 1));
-        }
-        if (acc.second != Integer.MIN_VALUE) {
-            out.collect(Tuple2.of(acc.second, 2));
-        }
-    }
-
-}
